+++ exporter/datadogexporter/traces_exporter_test.go
@@ -15,17 +15,16 @@
 package datadogexporter
 
 import (
-	"compress/gzip"
 	"context"
 	"encoding/json"
-	"io/ioutil"
+	"fmt"
 	"net/http"
 	"net/http/httptest"
+	"os"
 	"testing"
+	"time"
 
-	"github.com/DataDog/datadog-agent/pkg/trace/exportable/pb"
-	"github.com/DataDog/datadog-agent/pkg/trace/exportable/stats"
-	"github.com/gogo/protobuf/proto"
+	tracelog "github.com/DataDog/datadog-agent/pkg/trace/log"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"go.opentelemetry.io/collector/component/componenttest"
@@ -39,24 +38,82 @@ import (
 	"github.com/open-telemetry/opentelemetry-collector-contrib/exporter/datadogexporter/internal/testutils"
 )
 
-func testTracesExporterHelper(td ptrace.Traces, t *testing.T) []string {
+func TestMain(m *testing.M) {
+	tracelog.SetLogger(&testlogger{})
+	os.Exit(m.Run())
+}
+
+type testlogger struct{}
+
+// Trace implements Logger.
+func (testlogger) Trace(v ...interface{}) {}
+
+// Tracef implements Logger.
+func (testlogger) Tracef(format string, params ...interface{}) {}
+
+// Debug implements Logger.
+func (testlogger) Debug(v ...interface{}) { fmt.Println("DEBUG", fmt.Sprint(v...)) }
+
+// Debugf implements Logger.
+func (testlogger) Debugf(format string, params ...interface{}) {
+	fmt.Println("DEBUG", fmt.Sprintf(format, params...))
+}
+
+// Info implements Logger.
+func (testlogger) Info(v ...interface{}) { fmt.Println("INFO", fmt.Sprint(v...)) }
+
+// Infof implements Logger.
+func (testlogger) Infof(format string, params ...interface{}) {
+	fmt.Println("INFO", fmt.Sprintf(format, params...))
+}
+
+// Warn implements Logger.
+func (testlogger) Warn(v ...interface{}) error {
+	fmt.Println("WARN", fmt.Sprint(v...))
+	return nil
+}
+
+// Warnf implements Logger.
+func (testlogger) Warnf(format string, params ...interface{}) error {
+	fmt.Println("WARN", fmt.Sprintf(format, params...))
+	return nil
+}
+
+// Error implements Logger.
+func (testlogger) Error(v ...interface{}) error {
+	fmt.Println("ERROR", fmt.Sprint(v...))
+	return nil
+}
+
+// Errorf implements Logger.
+func (testlogger) Errorf(format string, params ...interface{}) error {
+	fmt.Println("ERROR", fmt.Sprintf(format, params...))
+	return nil
+}
+
+// Critical implements Logger.
+func (testlogger) Critical(v ...interface{}) error {
+	fmt.Println("CRITICAL", fmt.Sprint(v...))
+	return nil
+}
+
+// Criticalf implements Logger.
+func (testlogger) Criticalf(format string, params ...interface{}) error {
+	fmt.Println("CRITICAL", fmt.Sprintf(format, params...))
+	return nil
+}
+
+// Flush implements Logger.
+func (testlogger) Flush() {}
+
+func TestTraceExporter(t *testing.T) {
 	metricsServer := testutils.DatadogServerMock()
 	defer metricsServer.Close()
 
-	var got []string
+	var got string
 	server := httptest.NewServer(http.HandlerFunc(func(rw http.ResponseWriter, req *http.Request) {
 		assert.Equal(t, "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa", req.Header.Get("DD-Api-Key"))
-
-		contentType := req.Header.Get("Content-Type")
-
-		data := []string{contentType}
-		got = append(got, data...)
-
-		if contentType == "application/x-protobuf" {
-			testProtobufTracePayload(t, rw, req)
-		} else if contentType == "application/json" {
-			testJSONTraceStatsPayload(t, rw, req)
-		}
+		got = req.Header.Get("Content-Type")
 		rw.WriteHeader(http.StatusAccepted)
 	}))
 
@@ -87,72 +144,14 @@ func testTracesExporterHelper(td ptrace.Traces, t *testing.T) []string {
 	params := componenttest.NewNopExporterCreateSettings()
 	f := NewFactory()
 	exporter, err := f.CreateTracesExporter(context.Background(), params, &cfg)
-
 	assert.NoError(t, err)
 
-	defer func() {
-		assert.NoError(t, exporter.Shutdown(context.Background()))
-	}()
-
 	ctx := context.Background()
-	errConsume := exporter.ConsumeTraces(ctx, td)
-	assert.NoError(t, errConsume)
-
-	return got
-}
-
-func testProtobufTracePayload(t *testing.T, rw http.ResponseWriter, req *http.Request) {
-	var traceData pb.TracePayload
-	b, err := ioutil.ReadAll(req.Body)
-
-	if err != nil {
-		http.Error(rw, err.Error(), http.StatusInternalServerError)
-		assert.NoError(t, err, "http server received malformed trace payload")
-		return
-	}
-
-	defer req.Body.Close()
-
-	if marshallErr := proto.Unmarshal(b, &traceData); marshallErr != nil {
-		http.Error(rw, marshallErr.Error(), http.StatusInternalServerError)
-		assert.NoError(t, marshallErr, "http server received malformed trace payload")
-		return
-	}
-
-	assert.NotNil(t, traceData.Env)
-	assert.NotNil(t, traceData.HostName)
-	assert.NotNil(t, traceData.Traces)
-}
-
-func testJSONTraceStatsPayload(t *testing.T, rw http.ResponseWriter, req *http.Request) {
-	var statsData stats.Payload
-
-	gz, err := gzip.NewReader(req.Body)
-	if err != nil {
-		http.Error(rw, err.Error(), http.StatusInternalServerError)
-		require.NoError(t, err, "http server received malformed stats payload")
-		return
-	}
-
-	defer req.Body.Close()
-	defer gz.Close()
-
-	statsBytes, err := ioutil.ReadAll(gz)
-	if err != nil {
-		http.Error(rw, err.Error(), http.StatusInternalServerError)
-		require.NoError(t, err, "http server received malformed stats payload")
-		return
-	}
-
-	if marshallErr := json.Unmarshal(statsBytes, &statsData); marshallErr != nil {
-		http.Error(rw, marshallErr.Error(), http.StatusInternalServerError)
-		require.NoError(t, marshallErr, "http server received malformed stats payload")
-		return
-	}
-
-	assert.NotNil(t, statsData.Env)
-	assert.NotNil(t, statsData.HostName)
-	assert.NotNil(t, statsData.Stats)
+	err = exporter.ConsumeTraces(ctx, simpleTraces())
+	assert.NoError(t, err)
+	time.Sleep(10 * time.Millisecond) // we need a bit of time for channels to receive things
+	require.NoError(t, exporter.Shutdown(context.Background()))
+	require.Equal(t, "application/x-protobuf", got)
 }
 
 func TestNewTracesExporter(t *testing.T) {
@@ -211,23 +210,14 @@ func TestPushTraceData(t *testing.T) {
 	assert.Equal(t, recvMetadata.InternalHostname, "custom-hostname")
 }
 
-func TestTraceAndStatsExporter(t *testing.T) {
-	// ensure that the protobuf serialized traces payload contains HostName Env and Traces
-	// ensure that the json gzipped stats payload contains HostName Env and Stats
-	got := testTracesExporterHelper(simpleTraces(), t)
-
-	// ensure a protobuf and json payload are sent
-	assert.Equal(t, 2, len(got))
-	assert.Equal(t, "application/json", got[1])
-	assert.Equal(t, "application/x-protobuf", got[0])
-}
-
 func simpleTraces() ptrace.Traces {
-	return simpleTracesWithID(pcommon.NewTraceID([16]byte{1, 2, 3, 4}))
+	return simpleTracesWithID(pcommon.NewTraceID([16]byte{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 2, 3, 4}))
 }
 
 func simpleTracesWithID(traceID pcommon.TraceID) ptrace.Traces {
 	traces := ptrace.NewTraces()
-	traces.ResourceSpans().AppendEmpty().ScopeSpans().AppendEmpty().Spans().AppendEmpty().SetTraceID(traceID)
+	span := traces.ResourceSpans().AppendEmpty().ScopeSpans().AppendEmpty().Spans().AppendEmpty()
+	span.SetTraceID(traceID)
+	span.SetSpanID(pcommon.NewSpanID([8]byte{0, 0, 0, 0, 1, 2, 3, 4}))
 	return traces
 }