+++ CHANGELOG.md
@@ -51,6 +51,7 @@
 - `transformprocessor`: Add new `limit` function to allow limiting the number of items in a map, such as the number of attributes in `attributes` or `resource.attributes` (#9552)
 - `processor/attributes`: Support attributes set by server authenticator (#9420)
 - `datadogexporter`: Experimental support for Exponential Histograms with delta aggregation temporality (#8350)
+- `datadogexporter`: Unify traces exporter behavior with Datadog Agent OTLP traces ingest
 
 ### 🧰 Bug fixes 🧰